# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/readiness/{assignmentId}:
    get:
      operationId: getAssignmentReadiness
      tags: [analytics]
      summary: Surgically precise readiness pointers for a specific graded assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentReadiness"
...
components:
  schemas:
...
    AssignmentReadiness:
      type: object
      required: [assignmentId, readinessPercent, readyLabel, summary, recommendedPracticeRuns, perTopic]
      properties:
        assignmentId: { type: integer }
        readinessPercent: { type: number }
        readyLabel:
          type: string
          enum: [ready, almost, not_ready, untested]
        summary: { type: string }
        recommendedPracticeRuns: { type: integer }
        perTopic:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessTopic"
    ReadinessTopic:
      type: object
      required: [topicId, topicTitle, attempts, accuracy, pointer]
      properties:
        topicId: { type: integer }
        topicTitle: { type: string }
        attempts: { type: integer }
        accuracy: { type: number }
        masteryLabel: { type: ["string", "null"] }
        pointer: { type: string }
...