# lib/api-zod/src/generated/types/assignmentReadiness.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentReadinessReadyLabel } from './assignmentReadinessReadyLabel';
import type { ReadinessTopic } from './readinessTopic';

export interface AssignmentReadiness {
  assignmentId: number;
  readinessPercent: number;
  readyLabel: AssignmentReadinessReadyLabel;
  summary: string;
  recommendedPracticeRuns: number;
  perTopic: ReadinessTopic[];
}