# lib/api-zod/src/generated/types/readinessTopic.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReadinessTopic {
  topicId: number;
  topicTitle: string;
  attempts: number;
  accuracy: number;
  /** @nullable */
  masteryLabel?: string | null;
  pointer: string;
}